enable 1st and last cells without marks
now there is no need to add a marker "##" at the beginning or end
of the file to make a cell, the 1st cell can be defined as from
the beginning of file to the 1st ##, and the last cell can be from
the last ## to the end of file

This is achieved by modifying the `RunTmuxPythonCell()` function.

To search the cell end line:

```
  let l:cell_end = search(b:cellmode_cell_delimiter, 'W')
  if l:cell_end == 0
	  " if not found, assuming end of file
	  let l:cell_end = '$'
  else
	  " if found, move one line up
	  let l:cell_end -= 1
  endif
```

Similarly the cell_start line number is searched.
Then just yank the range:

```
:cell_start, cell_end y a
```

`cell_start` could be 1 (beginning of file, whether it is ## or not),
`cell_end` could be $ (end of file, whether it is ## or not).

Also in RunTmuxPythonAllCellsAbove(), a similar backward search is done.
If it found line 1, no cell is run.